Writing and refactoring tests
