+++ tox.ini
@@ -10,6 +10,8 @@ extras =
     test
 passenv =
     VK_ACCESS_TOKEN
+    VK_USER_LOGIN
+    VK_USER_PASSWORD
 commands =
     pytest {tty:--color=yes} {posargs: \
         --no-cov-on-fail --cov-report xml --cov-report term-missing --cov-append \
